src/tiktak/serialization/Error.java
package tiktak.serialization;

import java.io.IOException;
import java.util.Objects;

public class Error extends Message {

    private int code;
    private String message;

    @Override
    public void encode(MessageOutput out) throws IOException {

    }

    public Error(int code, String message){
        super();
    }


    @Override
    public String toString() {
        return "Error{" +
                "code=" + code +
                ", message='" + message + '\'' +
                '}';
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Error error = (Error) o;
        return code == error.code &&
                Objects.equals(message, error.message);
    }

    @Override
    public int hashCode() {
        return Objects.hash(code, message);
    }

    public int getCode() {
        return code;
    }

    public void setCode(int code) {
        this.code = code;
    }

    public String getMessage() {
        return message;
    }

    public void setMessage(String message) {
        this.message = message;
    }
}
